[Change] Move CryptlexClientFactory to the library itself
