+++ Examples/ConsoleAppNoGenericHostExample/ConsoleAppNoGenericHostExample.csproj
@@ -9,7 +9,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\Source\Cryptlex.Net\Cryptlex.Net.csproj" />
-    <ProjectReference Include="..\Common\Common.csproj" />
   </ItemGroup>
 
 </Project>
