+++ Examples/WpfAppExample/MainWindow.xaml.cs
@@ -1,4 +1,5 @@
 ﻿using Cryptlex.Net.Core;
+using Cryptlex.Net.Factories;
 using Cryptlex.Net.Licenses;
 using System.Linq;
 using System.Windows;
